chore(p2): move helper functions to new file
